chore: move lifecycle type to class
